Agregado los objetos a juntar y llevar a los pentagramas

+++ src/states/GameState.hx
@@ -12,6 +12,7 @@ import flixel.util.FlxRandom;
 import gameObjects.Enemy;
 import gameObjects.Pentagram;
 import gameObjects.Player;
+import gameObjects.RitualObject;
 import openfl.Assets;
 
 /**
@@ -31,6 +32,7 @@ class GameState extends FlxState
 	var amountOfPentagrams:Int = 4;
 	var pentagramsCollected:Int = 0;
 	var pentagrams:FlxTypedGroup<Pentagram>;
+	var ritualObjects:FlxTypedGroup<RitualObject>;
 	
 	//ENEMY VARS
 	var amountEnemies:Int=1;
@@ -55,13 +57,40 @@ class GameState extends FlxState
 		super.update();
 		FlxG.collide(map, player);
 		FlxG.collide(enemies, map, enemyMapCollide);
-		//FlxG.collide(enemies, map);
 		FlxG.collide(enemies, player, enemyPlayer);
 		FlxG.collide(enemies, null, enemyEnemyCollide);
 		
-		
+		checkRitualObjectsCollision();
 		checkPentagramsCollision();
 		checkWinCondition();
+	}
+	
+	private function checkRitualObjectsCollision() {
+		if (player.overlaps(ritualObjects)) {
+			var allRitualObjects:Array<RitualObject> = ritualObjects.members;
+			var ritualObject:RitualObject;
+			var done:Bool = false;
+			
+			while (!done) {
+				ritualObject = allRitualObjects.shift();
+				if (player.overlaps(ritualObject)) {
+					
+					if (player.ritualObjectHold != null) {
+						//REVIVO EL OBJETO VIEJO; PENSADO PARA PODER LEVANTAR OBJETOS DISTINTOS PARA CIRCULOS DISTINTOS
+						player.ritualObjectHold.x = player.x;
+						player.ritualObjectHold.y = player.y;
+						player.ritualObjectHold.revive();
+					}
+					//TOMO EL NUEVO OBJETO
+					player.ritualObjectHold = ritualObject;
+					ritualObject.kill();
+					
+					done = true;
+				}
+				//SIEMPRE LO DEVOLVEMOS AL ARRAY PORQUE SE DESTRUYEN AL LLEVARLO AL CIRCULO DEL RITUAL
+				allRitualObjects.push(ritualObject);
+			}
+		}
 	}
 	
 	private function initEnemies() {
@@ -95,15 +124,32 @@ class GameState extends FlxState
 			while (!done) {
 				pentagram = allPentagrams.shift();
 				pentaChecks++;
-				if (player.overlaps(pentagram)) {
+				if (player.overlaps(pentagram) && joinRitualObjectAndPentagram(pentagram)) {
 					done = true;
 					pentagram.kill();
 					pentagramsCollected++;
-					FlxG.log.advanced("Pentagram removed. Pentagrams checked: "+pentaChecks);
+					FlxG.log.advanced("Pentagram removed. Pentagrams checked: " + pentaChecks);
+					
+					//ME DESHAGO POR COMPLETO DEL OBJETO
+					ritualObjects.remove(player.ritualObjectHold);
+					player.ritualObjectHold.destroy();
+					player.ritualObjectHold = null; //NO SE SI SERÍA NECESARIO, PERO POR LAS DUDAS...
 				}else {
 					allPentagrams.push(pentagram);
+					if (pentaChecks > allPentagrams.length) {
+						FlxG.log.advanced("Revise " + pentaChecks + ". No puedo borrarlo");
+						done = true;
+					}
 				}
 			}
+		}
+	}
+	
+	public function joinRitualObjectAndPentagram(pentagram:Pentagram):Bool {
+		if (player.ritualObjectHold == null) {
+			return false;
+		}else {
+			return player.ritualObjectHold.getType() == pentagram.getType();
 		}
 	}
 	
@@ -119,6 +165,7 @@ class GameState extends FlxState
 	}
 	
 	private function initPentagrams() {
+		//PENTAGRAMS
 		pentagrams = new FlxTypedGroup<Pentagram>();
 		var pentagram:Pentagram;
 		
@@ -135,6 +182,24 @@ class GameState extends FlxState
 		pentagrams.add(pentagram);
 		
 		add(pentagrams);
+		
+		//RITUAL OBJECTS
+		ritualObjects = new FlxTypedGroup<RitualObject>();
+		var ritualObject:RitualObject;
+		
+		ritualObject = new RitualObject(14 * tileSize, 8 * tileSize);
+		ritualObjects.add(ritualObject);
+		
+		ritualObject = new RitualObject(17 * tileSize, 8 * tileSize);
+		ritualObjects.add(ritualObject);
+		
+		ritualObject = new RitualObject(14 * tileSize, 11 * tileSize);
+		ritualObjects.add(ritualObject);
+		
+		ritualObject = new RitualObject(17 * tileSize, 11 * tileSize);
+		ritualObjects.add(ritualObject);
+		
+		add(ritualObjects);
 	}
 	
 	private function enemyMapCollide(enemy:Enemy, map:FlxTilemap):Void {
